+++ examples/resources/dokploy_tag/import.sh
@@ -1,0 +1,1 @@
+terraform import dokploy_tag.production LPSgSJ1Jyxlu0CvEnOfyp